[FIX] issues with etcd download on docker-hub

+++ container/build.sh
@@ -21,7 +21,7 @@ ETCD_VER=v3.4.14
 addgroup --gid 1020 etcd
 
 adduser --system --home /var/lib/etcd --gid 1020 --uid 1020 --quiet etcd
-adduser odoo syslog
+adduser etcd syslog
 
 mkdir -p /home/etcd
 mkdir -p /container/data
@@ -38,11 +38,23 @@ rm -f /tmp/etcd-${ETCD_VER}-linux-amd64.tar.gz
 rm -rf /tmp/etcd-download-test && mkdir -p /tmp/etcd-download-test
 
 curl -L ${DOWNLOAD_URL}/${ETCD_VERSION}/etcd-${ETCD_VERSION}-linux-amd64.tar.gz -o /tmp/etcd-${ETCD_VERSION}-linux-amd64.tar.gz
+if [ $? -ne 0 ]; then
+    exit -1
+fi
 tar xzvf /tmp/etcd-${ETCD_VERSION}-linux-amd64.tar.gz -C /usr/bin --strip-components=1
+if [ $? -ne 0 ]; then
+    exit -1
+fi
 rm -f /tmp/etcd-${ETCD_VERSION}-linux-amd64.tar.gz
 
 /usr/bin/etcd --version
+if [ $? -ne 0 ]; then
+    exit -1
+fi
 /usr/bin/etcdctl version
+if [ $? -ne 0 ]; then
+    exit -1
+fi
 
 #--
 # Cleaning
